+++ sparklines/lib/data/data_point.dart
@@ -1,6 +1,7 @@
 import 'dart:ui' show lerpDouble;
 import 'dart:math' as math;
 
+import 'package:any_sparklines/interfaces/chart_border.dart';
 import 'package:any_sparklines/interfaces/pie_offset.dart';
 
 import '../interfaces/data_point_data.dart';
@@ -90,6 +91,7 @@ extension DataPointExtension on DataPoint {
   IThicknessOverride? get thickness => of<IThicknessOverride>();
   IDataPointStyle? get style => of<IDataPointStyle>();
   IPieOffset? get pieOffset => of<IPieOffset>();
+  IDataPointBorder? get border => of<IDataPointBorder>();
 
 }