+++ sparklines/lib/interfaces/chart_area.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+abstract class IChartArea {
+  Gradient? get areaGradient;
+  Color? get areaColor;
+  PathFillType? get areaFillType;
+}
+