+++ sparklines/lib/interfaces/data_point_style.dart
@@ -20,3 +20,7 @@ abstract class IDataPointRenderer {
     Object dataPoint,
   );
 }
+
+abstract class IChartDataPointStyle {
+  IDataPointStyle? get pointStyle;
+}